move deleteCarRequest to logic/carRequest.js
